requirements.txt
apns==1.1.2
boto==2.13.3
https://www.djangoproject.com/download/1.6b4/tarball/
-e git+https://github.com/dcramer/django-uuidfield.git@643ccc2ccbe8dfa2e073adbae68c1eac2eb70a50#egg=django-uuidfield
djangorestframework==2.3.7
-e git+https://github.com/dmeehan/django-rest-framework-gis.git@9ade355885f3cff07557749a5cf8ea5df90c9280#egg=djangorestframework-gis
django-storages==1.1.8
dj_database_url
eventlet==0.14.0
gunicorn==0.17.4
mock==1.0.1
psycopg2
raven>=3
south==0.8.1
